+++ database/migrations/2024_06_14_161810_create_transaction_items_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->foreignId('transaction_id')->constrained('buy_history');
             $table->foreignId('product_id')->constrained('products');
-            $table->bigInteger('price');
+            $table->double('price');
             $table->integer('quantity');
             $table->timestamps();
         });
